requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
backoff==1.11.1
certifi==2024.6.2
charset-normalizer==3.3.2
coverage==5.5
deprecation==2.1.0
eyed3==0.9.7
ffmpeg-python==0.2.0
filetype==1.2.0
frozenlist==1.4.1
future==1.0.0
idna==3.7
multidict==6.0.5
packaging==24.1
pytube==15.0.0
redis==5.0.5
requests==2.32.3
spotipy==2.24.0
toml==0.10.2
urllib3==2.2.1
yarl==1.9.4
ytmusicapi==1.7.3
